Move nginx-related functions from `addons.go` to `nginxingress.go`
